# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "🔱 atlantis API",
    "version": "1.0.0",
    "description": "API documentation for the Atlantis platform - diagrams, notes, and knowledge management."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Diagrams",
      "description": "See and manage your Mermaid diagrams"
    },
    {
      "name": "Notes",
      "description": "See and manage your notes"
    }
  ],
  "paths": {
    "/api/access/diagrams": {
      "get": {
        "tags": [
          "Diagrams"
        ],
        "summary": "List all diagrams",
        "description": "Retrieve a paginated list of all stored diagrams. Requires `ENABLE_API_ACCESS=true`.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number to retrieve"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of diagrams",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagramList"
                }
              }
            }
          },
          "403": {
            "description": "API Access Disabled"
          }
        }
      },
      "post": {
        "tags": [
          "Diagrams"
        ],
        "summary": "Create a new diagram",
        "description": "Create a new diagram with Mermaid content. Validates syntax before saving. Requires `ENABLE_API_ACCESS=true`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiagramRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Diagram created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or Mermaid syntax error"
          },
          "403": {
            "description": "API Access Disabled"
          }
        }
      }
    },
    "/api/access/diagrams/{id}": {
      "get": {
        "tags": [
          "Diagrams"
        ],
        "summary": "Get a single diagram",
        "description": "Retrieve the full details of a specific diagram by ID. Requires `ENABLE_API_ACCESS=true`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the diagram"
          }
        ],
        "responses": {
          "200": {
            "description": "The diagram details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Diagram"
                }
              }
            }
          },
          "404": {
            "description": "Diagram not found"
          },
          "403": {
            "description": "API Access Disabled"
          }
        }
      }
    },
    "/api/access/notes": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "List all notes",
        "description": "Retrieve a list of notes. Content is omitted from the list for performance. Requires `ENABLE_API_ACCESS=true`.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of notes to return"
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number to retrieve"
          }
        ],
        "responses": {
          "200": {
            "description": "List of notes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteList"
                }
              }
            }
          },
          "403": {
            "description": "API Access Disabled"
          }
        }
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "summary": "Create a new note",
        "description": "Create a new note. Requires `ENABLE_API_ACCESS=true`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNoteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "403": {
            "description": "API Access Disabled"
          }
        }
      }
    },
    "/api/access/notes/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "summary": "Get a note",
        "description": "Retrieve a specific note by ID. Requires `ENABLE_API_ACCESS=true`.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Note ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Note details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "404": {
            "description": "Note not found"
          },
          "403": {
            "description": "API Access Disabled"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Diagram": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "abc123"
          },
          "title": {
            "type": "string",
            "example": "System Architecture"
          },
          "content": {
            "type": "string",
            "example": "graph TD; A-->B;"
          },
          "emoji": {
            "type": "string",
            "example": "📊"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isFavorite": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CreateDiagramRequest": {
        "type": "object",
        "required": [
          "content"
        ],
        "properties": {
          "title": {
            "type": "string",
            "example": "New Diagram"
          },
          "content": {
            "type": "string",
            "example": "graph TD; A-->B;"
          }
        }
      },
      "DiagramList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                }
              }
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      },
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "note_123"
          },
          "title": {
            "type": "string",
            "example": "Meeting Notes"
          },
          "content": {
            "type": "string",
            "example": "# Agenda\n- Discuss project"
          },
          "language": {
            "type": "string",
            "example": "markdown"
          },
          "starred": {
            "type": "boolean",
            "example": true
          },
          "private": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateNoteRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Untitled Note"
          },
          "content": {
            "type": "string",
            "example": "Note content..."
          },
          "language": {
            "type": "string",
            "example": "markdown",
            "description": "Language or file type (e.g., markdown, todo, javascript)"
          }
        }
      },
      "NoteList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Note"
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer"
              },
              "limit": {
                "type": "integer"
              },
              "total": {
                "type": "integer"
              },
              "totalPages": {
                "type": "integer"
              }
            }
          }
        }
      }
    }
  }
}